cleaned up, added common functions to service
