+++ src/app/container/descriptors/descriptors.component.ts
@@ -35,7 +35,7 @@ import { FileService } from '../../shared/file.service';
 export class DescriptorsComponent extends FileSelector implements AfterViewChecked {
 
   @Input() id: number;
-  @Input() toolpath: string;
+  @Input() entrypath: string;
   content: string;
   contentHighlighted: boolean;
   constructor(private containerService: ContainerService,
@@ -77,17 +77,7 @@ export class DescriptorsComponent extends FileSelector implements AfterViewCheck
   }
 
   getDescriptorPath(descType) : string {
-    if (this.currentFile != null) {
-        const basepath = Dockstore.API_URI + '/api/ga4gh/v1/tools/';
-        let descriptor = 'plain-CWL';
-        if (descType === 'wdl') {
-          descriptor = 'plain-WDL';
-        }
-        let customPath = encodeURIComponent(this.toolpath) + '/versions/' + this.currentVersion.name + '/' + descriptor + '/descriptor/' + encodeURIComponent(this.currentFile.path);
-        return basepath + customPath;
-      } else {
-        return null;
-      }
-    }
+    return this.fileService.getDescriptorPath(this.entrypath, this.currentVersion, this.currentFile, this.currentDescriptor, 'tool');
+  }
 
 }
