+++ src/app/shared/file.service.ts
@@ -13,11 +13,55 @@
  *    See the License for the specific language governing permissions and
  *    limitations under the License.
  */
+ import { Dockstore } from './dockstore.model';
 
 export class FileService {
 
     /* Highlight Code */
     highlightCode(code: string): string {
       return '<pre><code class="yaml highlight">' + code + '</pre></code>';
+    }
+
+    getDescriptorPath(entrypath, currentVersion, currentFile, descriptor, entrytype) : string {
+      if (currentFile != null) {
+        const basepath = Dockstore.API_URI + '/api/ga4gh/v1/tools/';
+        let descriptorType = 'plain-CWL';
+        if (descriptor === 'wdl') {
+          descriptorType = 'plain-WDL';
+        }
+
+        let entry = '';
+        if (entrytype === 'workflow') {
+          entry = encodeURIComponent('#workflow/' + entrypath);
+        } else {
+          entry = encodeURIComponent(entrypath)
+        }
+
+        let customPath =  entry + '/versions/' + currentVersion.name + '/' + descriptorType + '/descriptor/' + encodeURIComponent(currentFile.path);
+        return basepath + customPath;
+      } else {
+        return null;
+      }
+    }
+
+    // Downloads a file
+    // TODO: Temporary, need to update test param endpoint to return raw file based on path
+    downloadFile(file, id) : void {
+      let filename = "dockstore.txt";
+      if (file != null) {
+        const splitFileName = (file.path).split('/');
+        filename = splitFileName[splitFileName.length - 1];
+      }
+      const data = "data:text/plain," + encodeURIComponent(file.content);
+
+      $('#' + id).attr('href', data).attr('download', filename);
+    }
+
+    // Get the path of the file
+    getFilePath(file) : string {
+      if (file != null) {
+        return file.path;
+      }
+      return null;
     }
 }
