+++ src/app/workflow/files/files.component.html
@@ -16,11 +16,11 @@
 
 <tabset [justified]="true">
   <tab heading="Descriptor Files">
-    <app-descriptors-workflow [toolpath]="toolpath" [id]="id" [versions]="versions" [default]="default">
+    <app-descriptors-workflow [entrypath]="entrypath" [id]="id" [versions]="versions" [default]="default">
     </app-descriptors-workflow>
   </tab>
   <tab heading="Test Parameter Files">
-    <app-paramfiles-workflow [toolpath]="toolpath" [id]="id" [versions]="versionsWithParamfiles" [default]="default">
+    <app-paramfiles-workflow [entrypath]="entrypath" [id]="id" [versions]="versionsWithParamfiles" [default]="default">
     </app-paramfiles-workflow>
   </tab>
 </tabset>
