+++ src/app/workflow/paramfiles/paramfiles.component.ts
@@ -30,7 +30,7 @@ import { Dockstore } from '../../shared/dockstore.model';
 })
 export class ParamfilesWorkflowComponent extends FileSelector implements AfterViewChecked {
   @Input() id: number;
-  @Input() toolpath: string;
+  @Input() entrypath: string;
   content: string;
 
   contentHighlighted: boolean;
@@ -72,23 +72,12 @@ export class ParamfilesWorkflowComponent extends FileSelector implements AfterVi
   }
 
   // Downloads a file
-  // TODO: Temporary, need to update test param endpoint to return raw file based on path
   downloadFile(file, id) : void {
-    let filename = "dockstore.txt";
-    if (file != null) {
-      const splitFileName = (file.path).split('/');
-      filename = splitFileName[splitFileName.length - 1];
-    }
-    const data = "data:text/plain," + encodeURIComponent(file.content);
-
-    $('#' + id).attr('href', data).attr('download', filename);
+    this.fileService.downloadFile(file, id);
   }
 
   // Get the path of the file
   getFilePath(file) : string {
-    if (file != null) {
-      return file.path;
-    }
-    return null;
+    return this.fileService.getFilePath(file);
   }
 }
