+++ src/styles.scss
@@ -1038,6 +1038,10 @@ div.tabs>div:target>div,
   position: absolute;
 }
 
+.code-copy-button > a {
+  border: 1px solid transparent !important;
+}
+
 .ellipsis-lines {
   white-space: nowrap;
   text-overflow: ellipsis;
